+++ packages/core-components/src/lib/unsorted/ui/Formatting/CurrencyFormatGrid.svelte
@@ -22,7 +22,7 @@
 	<div transition:slide>
 		<table>
 			<thead>
-				<th class="align_left narrow_column">Format Tag</th>
+				<th class="align_left narrow_column">Format Name</th>
 				<th class="align_left wide_column">Format Code</th>
 				<th class="align_left wide_column">Example Input</th>
 				<th class="align_right wide_column">Example Output</th>